% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_mod1_sims.R
\name{agg_mod1_sims}
\alias{agg_mod1_sims}
\title{Aggregate Model 1 Simulations}
\usage{
agg_mod1_sims(sims)
}
\arguments{
\item{sims}{data.frame containing simulations from model 3.}
}
\value{
A data.frame with columns: \itemize{
\item{time: }{Numerical year}
\item{q05: }{2.5 percentile of reported cholera cases}
\item{mean: }{Mean of reported cholera cases}
\item{q50: }{Median of reported cholera cases}
\item{q95: }{97.5 percentile of reported cholera cases}
}
}
\description{
This function summarizes simulations from model 1.
}
