% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_mod3_sims.R
\name{agg_mod3_sims}
\alias{agg_mod3_sims}
\title{Aggregate Model 3 Simulations}
\usage{
agg_mod3_sims(sims)
}
\arguments{
\item{sims}{data.frame containing simulations from model 3.}
}
\value{
A data.frame with columns: \itemize{
\item{time: }{Numerical year}
\item{q05: }{2.5 percentile of reported cholera cases}
\item{mean: }{Mean of reported cholera cases}
\item{q50: }{Median of reported cholera cases}
\item{q95: }{97.5 percentile of reported cholera cases}
}
}
\description{
Because Model 3 has spatial structure, simulations from model 3 result in
simulations of the process for each departement. For comparitive purposes,
it's useful to aggregate these results into national counts, and this
function does precisly that.
}
