% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1_NegBinom.R
\name{ar1_NegBinom}
\alias{ar1_NegBinom}
\title{Autoregressive Negative Binomial}
\usage{
ar1_NegBinom(
  data,
  init,
  start_vals = c(10, 0.5, 1),
  transform = TRUE,
  a = 2,
  ...
)
}
\arguments{
\item{data}{The data to fit, numeric vector.}

\item{init}{Initial value}

\item{start_vals}{Starting values for the parameters}

\item{transform}{Boolean, whether or not to use variable transformations in the optimization}

\item{a}{Only used if \code{transform == TRUE}. \code{a} is a numeric representing the maximum value
for the AR1 coefficient, and \code{-a} is the minimum value.}

\item{...}{other parameters to be passed into the \code{optim} function}
}
\value{
An AR1_NegBinom model
}
\description{
Fits a simple AR1 negative binomial model. Let \eqn{Y_1, \ldots, Y_N} be
the \eqn{N} random variables representing observed data
\eqn{y_1^*, \ldots y_N^*}, respectively. The model assumes that
\eqn{Y_n|Y_{n-1}} for some \eqn{n \in 1:N} has the following distribution:
\deqn{Y_n | Y_{n - 1} = y_{n-1}* \sim NB(a + by_{n-1}^*, \psi)}
Where \eqn{a, b} and \eqn{\psi} are model parameters that need to be
estimated.
}
\details{
Missing values are currently ignored.
}
\examples{
ar1_NegBinom(round(sunspot.year), init = 5)
}
