% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti1_covars.R
\name{covars}
\alias{covars}
\title{Make covariate table for Model 1}
\usage{
covars(
  tmin,
  tmax,
  byt = 1,
  nbasis = 6,
  degree = 6,
  per = 52.14,
  data,
  settings
)
}
\arguments{
\item{tmin}{first observation time}

\item{tmax}{last observation time}

\item{byt}{number of steps between each time}

\item{nbasis}{number of basis spline terms}

\item{degree}{degrees of freedom for the basis splines}

\item{settings}{settings for the vaccination campaign scenario}
}
\description{
Generate a data frame containing weekly values for periodic basis splines for the seasonal transmission term beta.
}
