% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_logLik2.R
\name{est_logLik2}
\alias{est_logLik2}
\title{Estimate Log-Likelihood}
\usage{
est_logLik2(model)
}
\arguments{
\item{model}{Input is a spatPomp object of the model whose Log-likelihood
is to be estimated. The spatPomp object should already have parameter values
stored in it.}
}
\value{
The negative log likelihood of the model.
}
\description{
This function estimates the log-likelihood for model 2.
}
