% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_logLik3.R
\name{est_logLik3}
\alias{est_logLik3}
\title{Estimate Log-Likelihood}
\usage{
est_logLik3(
  version = "panelPomp",
  params = NULL,
  Np = 100,
  nreps = 10,
  ncores = 1
)
}
\arguments{
\item{version}{in c("panelPomp", "spatPomp", "pomp").}

\item{params}{Parameters that will be used in the log-likelihood estimation.
Note that these parameters need to match the version of the model that
will be used, otherwise a warning will be given and the default parameters
for the version requested will be used.}

\item{Np}{Number of Particles used in the (potentially block) Particle Filter}

\item{nreps}{Number of repeated particle filters}

\item{ncores}{Number of cores used to conduct the repeated particle filters.
The parallelization is done over \code{nreps}, so ideally \code{ncores} should divide
\code{nreps}.}
}
\description{
This function estimates the log-likelihood for model 3.
}
\details{
As of Feb 11, 2022, Model 3 can be written in 3 different forms:
\itemize{
\item Coupled \code{pomp} object (\code{haiti3_correct()}).
\item Independent \code{panelPomp} object (\code{haiti3_panel()}).
\item Coupled \code{spatPomp} object (\code{haiti3_spatPomp()}).
}
}
