% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_haiti1.R
\name{fit_haiti1}
\alias{fit_haiti1}
\title{Fit Haiti 1}
\usage{
fit_haiti1(start_params, NMIF, NP, NP_EVAL, NREPS_EVAL, RW_SD, COOLING = 0.5)
}
\arguments{
\item{start_params}{A numeric vector of parameters for Model 1.}

\item{NMIF}{Integer number of IF2 iterations.}

\item{NP}{Integer number of particles for the IF2 algorithm.}

\item{NP_EVAL}{Integer number of particles to use in the particle filter
that is used to estimate the log-likelihood of the fitted model.}

\item{NREPS_EVAL}{Number of repeated likelihood evaluations.}

\item{RW_SD}{Random walk standard deviations for the IF2 algorithm.}

\item{COOLING}{Cooling rate for the IF2 algorithm.}
}
\value{
A numeric vector containing the log-likelihood estimate of the model
parameters, an estimate of the standard error of the estimated likelihood,
and the parameter vector resulting from the IF2 algorithm.
}
\description{
This simple function takes as input starting values for Model 1 parameters,
as well as hyperparameters to the IF2 algorithm and fits Model 1 using these
input values.
}
