% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_haiti1_SudEst.R
\name{fit_haiti1_SudEst}
\alias{fit_haiti1_SudEst}
\title{Fit Model 1 - Sud Est}
\usage{
fit_haiti1_SudEst(
  NP = 50,
  NMIF = 3,
  NSEQ = 3,
  NREPS_EVAL = 3,
  NP_EVAL = 50,
  ncores = 3
)
}
\arguments{
\item{NP}{Number of particles to use for the Iterated Filter}

\item{NMIF}{Number of MIF iterations}

\item{NSEQ}{Number of global search starting points.}

\item{NREPS_EVAL}{Number of particle filter replications to evaluate the
likelihood.}

\item{NP_EVAL}{Number of particles used for likelihood evaluation.}

\item{ncores}{Number of cores to use for parallelization.}
}
\description{
This function fits Model 1 by performing calculating a Monte-Carlo adjusted
profile for the linear trend in transmission parameter.
}
\examples{
fit_m1 <- fit_haiti1_SudEst()

}
