% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_haiti3.R
\name{fit_haiti3}
\alias{fit_haiti3}
\title{Fit Haiti3}
\usage{
fit_haiti3(
  start_params,
  NBPF = 5,
  NP = 50,
  SPAT_REGRESSION = 0.05,
  NP_EVAL = 100,
  NREPS_EVAL = 6,
  RW_SD = NULL,
  COOLING = 0.5,
  start_date = "2010-11-20"
)
}
\arguments{
\item{start_params}{a numeric vector of parameters for Haiti 3 model.}

\item{NBPF}{number of BPF iterations to use for the IBPF algorithms.}

\item{NP}{Number of particles for the IBPF algorithm.}

\item{SPAT_REGRESSION}{The regression parameter for shared parameters in the
IBPF algorithm.}

\item{NP_EVAL}{Number of particles to use in the evaluation of parameter estimates.}

\item{NREPS_EVAL}{Number of replicates of the BPF likelihood evaluation.}

\item{RW_SD}{Random Walk SD for the IBPF algorithm.}

\item{COOLING}{Cooling for the IBPF algorithm}

\item{start_date}{Starting date used for Haiti 3.}
}
\value{
a numeric vector containing the BPF likelihood estimate of the
calibrated parameters, the standard error of this estimate, followed by
all of the parameters of the model (both estimated and fixed parameters.)
}
\description{
A simple function that takes as input starting parameters, as well as
hyperparameters to the IBPF algorithm and fits haiti3 using the given
starting parameters.
}
\examples{
\dontrun{fit_m1 <- fit_haiti1()}
}
