% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elimProbs.R
\name{get_elimProbs}
\alias{get_elimProbs}
\title{Get Elimination Probabilities}
\usage{
get_elimProbs(sims, model)
}
\arguments{
\item{model}{}
}
\value{
A list containing the Probability of Elimination (probElim),
Cumulative Infections (cumInf), and Probability of Elimination over time
(ElimTime).
}
\description{
This function takes simulations from a model as input, and outputs a list
containing the probability of elimination, Cumulative Infections,
and probability of elimination over time.
}
\details{
The probability of elimination is calculated as:
\itemize{
\item{Probability of Elimination}{The proportion of simulations that
achieved less than one infection with V \emph{cholerae} (including reported
and unreported infections) over at least 52 consecutive weeks.}
\item{Cumulative Infections}{Median estimate with 95\% CI of cumulative
infections from February 2019, to February 2024.}
\item{Probability of elimination over time}{he proportion of simulations that
achieved less than one infection with V \emph{cholerae} (including reported
and unreported infections) over at least 52 consecutive weeks, over time.}
}
}
