% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_h1_seasonality.R
\name{get_h1_seasonality}
\alias{get_h1_seasonality}
\title{Get Haiti 1 Seasonality}
\usage{
get_h1_seasonality(mod1, times = 431:800)
}
\arguments{
\item{mod1}{a pomp object representing Model 1. It must contain parameters.}

\item{times}{a vector containing the weeks that will be used to estimate
the seasonality. By default, this is chosen as 431:800, which starts the
week after available training data and covers at least the entirety of
the following year. In this case, the linear trend in beta ends after the
available training data.}
}
\value{
data.frame containing three columns: (1) week, which is week number
corresponding to the measurement time (2) date, which is the date of the
measurement and (3) trans_std, which contains the value of the Model 1
seasonality component.
}
\description{
This function takes as input a pomp representation of Model 1 that contains
estimated parameter values, and returns a data.frame object with the
estimated seasonality over time.
}
