% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_h1_startVals.R
\name{get_h1_startVals}
\alias{get_h1_startVals}
\title{Get starting values for fitting Model 1}
\usage{
get_h1_startVals(n_params)
}
\arguments{
\item{n_params}{number of starting parameters to use.}
}
\value{
data.frame containing
}
\description{
This function get a set of starting parameters that are intended to
reproduce the results of Lee et al (2020). The starting parameter values
were not provided, so this function attempts to estimate them based on the
description of their distributions provided in the supplement material.
}
