% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_h3_S_quants.R
\name{get_h3_S_quants}
\alias{get_h3_S_quants}
\title{Get Susceptible Quantiles for Model 3}
\usage{
get_h3_S_quants(bpf, i)
}
\arguments{
\item{bpf}{A bpfiltered object with saved states.}

\item{i}{the time point \eqn{t_i} to get Susceptible qunatiles.}
}
\value{
A data.frame containing quantiles of the susceptible compartment at
time t_i.
}
\description{
This function gets quantiles for the Susceptible compartment from the saved
states of a block particle filter of Model 3. This function is only used in
the supplement material.
}
