% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model3_vacc_scenario_params.R
\name{get_model3_vacc_scenario_params}
\alias{get_model3_vacc_scenario_params}
\title{Model3 Vaccination Scenario Parameters}
\usage{
get_model3_vacc_scenario_params(
  scenario = c("noVacc", "2dep", "3dep", "slowNation", "fastNation", "highFastNation")
)
}
\arguments{
\item{scenario}{Vaccination scenario that is to be ran.}
}
\description{
This function returns the parameters that are necessary to run the vaccination
scenarios (in spatPomp form).
}
