% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vacc_scenario_params3.R
\name{get_vacc_scenario_params3}
\alias{get_vacc_scenario_params3}
\title{Get vaccination scenario parameters}
\usage{
get_vacc_scenario_params3(scenario_str = "S1")
}
\arguments{
\item{scenario_str}{character vector containing the name of each vaccination
scenario that will be simulated. The options are:
TODO: update this list
\itemize{
\item S0: Default vaccination scenario
\item S1
\item S2
\item S3
\item S4
\item S25
}}
}
\description{
This function returns the parameters that are needed to run the various
vaccination scenarios for model 3.
}
