% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choleraData.R
\docType{data}
\name{h1LeeStartsEpi}
\alias{h1LeeStartsEpi}
\title{Lee et al (2020) Model 1 Epidemic Parameters}
\format{
The dataset has 152 rows and 23 columns. Each column corresponds to
a parameter value for the epidemic version of Model 1, except for the column
\code{parid} which identifies the parameter set.
}
\source{
The data was created by running the source code of Lee et al (2020).
Because of differences in package versions and other possible numeric
considerations, the set of parameters are not the exact same as those
used by Lee et al (2020), but the same code was used to generate these
data, resulting in very similar sets of parameters. After generating the
data using the Lee et al (2020) souce code, some parameters with low log
likelihoods were removed, as well as other outlying values that were
inconsistent with those shown in the Lee et al (2020) Figures S8 and S9.
}
\usage{
h1LeeStartsEpi
}
\description{
This dataset contains parameters similar to those used by Lee et al (2020)
to project cholera incidence in Haiti using the epidemic version of Model 1.
}
\keyword{datasets}
