% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti1.R, R/haiti1_panel.R
\name{haiti1}
\alias{haiti1}
\title{Build pomp object for Model 1}
\usage{
haiti1()

haiti1()
}
\arguments{
\item{depts}{Number of departements in model}

\item{vacscen}{ID code for vaccination scenario}
}
\value{
An object of class \sQuote{pomp}

An object of class \sQuote{panelPomp}.
}
\description{
Generate a class \sQuote{pomp} object for fitting to epidemic/endemic Haiti cholera data

Generate a class \sQuote{panelPomp} object for fitting to epidemic/endemic Haiti cholera data.
}
\examples{
m1 <- haiti1()
m1 <- haiti1()
}
