% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti1_agg_data.R
\name{haiti1_agg_data}
\alias{haiti1_agg_data}
\title{Clean haiti cases data for Model 1}
\usage{
haiti1_agg_data()
}
\value{
A data frame
}
\description{
Generate a data frame of the reported cases of cholera in Haiti from October 2010 to
January 2019, aggregated over all departments.
}
\details{
Code adapted from Model 1 Team at JHU Bloomberg School of Public Health
}
\examples{
haiti1_agg_data()
}
