% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti1_data.R
\name{haiti1_data}
\alias{haiti1_data}
\title{Clean haiti cases data for Model 1}
\usage{
haiti1_data()
}
\value{
A data frame
}
\description{
Generate a data frame of the reported cases of cholera in Haiti from October 2010 to
January 2019, separated by departement.
}
\details{
Code adapted from Model 1 Team at JHU Bloomberg School of Public Health
}
\examples{
haiti1_data()
}
