% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti1_disagg.R
\name{haiti1_disagg}
\alias{haiti1_disagg}
\title{Build pomp object for Model 1 - Disaggregated Joint Model}
\usage{
haiti1_disagg()
}
\arguments{
\item{vacscen}{ID code for vaccination scenario}
}
\value{
An object of class \sQuote{pomp}
}
\description{
Generate a class \sQuote{pomp} object for fitting to epidemic/endemic Haiti cholera data jointly with separate departments.
}
\examples{
m1 <- haiti1_disagg(vacscen = "id0")
}
