% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti1_fit.R
\name{haiti1_fit}
\alias{haiti1_fit}
\title{Build pomp object for Model 1}
\usage{
haiti1_fit(period)
}
\arguments{
\item{period}{An integer to select the fitting period for the pomp model.
(0 = full (weeks 1-426, from October 2010 through January 2019),
1 = epidemic (weeks 1-232, from October 2010 through March 2015),
2 = endemic (weeks 233-426, from April 2015 through January 2019))}
}
\value{
An object of class \sQuote{pomp}.
}
\description{
Generate a class \sQuote{pomp} object for fitting to epidemic/endemic Haiti cholera data.
}
\examples{
m1 <- haiti1_fit(period = 1)
}
