% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti1_joint.R
\name{haiti1_joint}
\alias{haiti1_joint}
\title{Build pomp object for Model 1 - Joint}
\usage{
haiti1_joint(
  vacscen = "id0",
  breakpoint = 232,
  rho_flag = FALSE,
  tau_flag = TRUE,
  sig_sq_flag = TRUE,
  beta_flag = FALSE,
  nu_flag = FALSE
)
}
\arguments{
\item{vacscen}{ID code for vaccination scenario}

\item{breakpoint}{the week number at which to start re-estimation}

\item{rho_flag}{boolean indicating whether to re-estimate \eqn{\rho} at the breakpoint}

\item{tau_flag}{boolean indicating whether to re-estimate \eqn{\tau} at the breakpoint}

\item{sig_sq_flag}{boolean indicating whether to re-estimate \eqn{\sigma^2} at the breakpoint}

\item{beta_flag}{boolean indicating whether to re-estimate \eqn{\beta's} at the breakpoint}

\item{nu_flag}{boolean indicating whether to re-estimate \eqn{\nu} at the breakpoint}
}
\value{
An object of class \sQuote{pomp}
}
\description{
Generate a class \sQuote{pomp} object for fitting to epidemic/endemic Haiti cholera data jointly.
}
\examples{
\dontrun{
m1 <- haiti1_joint(
  vacscen = "id0", breakpoint = 232, rho_flag = T,
  tau_flag = F, sig_sq_flag = T, beta_flag = F, nu_flag = F
  )
}
}
