% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti1_panel.R
\name{haiti1_panel}
\alias{haiti1_panel}
\title{Build panelPomp object for Model 1}
\usage{
haiti1_panel(vacscen = "id0")
}
\arguments{
\item{vacscen}{Vaccination scenario}
}
\value{
An object of class \sQuote{panelPomp}.
}
\description{
Generate a class \sQuote{panelPomp} object for fitting to epidemic/endemic Haiti cholera data.
}
\examples{
m1 <- haiti1_panel(vacscen = 'id0')
}
