% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti1_supp_params.R
\docType{data}
\name{haiti1_supp_params}
\alias{haiti1_supp_params}
\title{Reported estimated parameters from Model 1 supplement}
\format{
A data frame with 21 rows and 2 columns, one for each period.
\describe{
\item{rho}{reporting rate}
\item{tau}{observation process dispersion}
\item{beta1}{seasonal transmission parameter (1st component of basis spline)}
\item{beta2}{seasonal transmission parameter (2nd component of basis spline)}
\item{beta3}{seasonal transmission parameter (3rd component of basis spline)}
\item{beta4}{seasonal transmission parameter (4th component of basis spline)}
\item{beta5}{seasonal transmission parameter (5th component of basis spline)}
\item{beta6}{seasonal transmission parameter (6th component of basis spline)}
\item{sigma}{latent period}
\item{theta0}{proportion of non-vaccinated Exposed becoming Asymptomatic}
\item{alpha}{mean duration of natural immunity}
\item{mu}{birth rate}
\item{delta}{death rate}
\item{nu}{mixing coefficient}
\item{S_0}{initial proportion of Susceptible at the start of the period}
\item{E_0}{initial proportion of Exposed at the start of the period}
\item{I_0}{initial proportion of Infectious at the start of the period}
\item{A_0}{initial proportion of Asymptomatic at the start of the period}
\item{R_0}{initial proportion of Recovered at the start of the period}
\item{pop_0}{initial population at the start of the period}
}
}
\source{
\url{https://zenodo.org/record/3360991#.YNC2Iy1h0ow}
}
\usage{
data(haiti1_supp_params)
}
\description{
Dataset containing the estimated parameters sets after fitting to the epidemic and endemic periods.
}
\references{

}
\keyword{datasets}
