% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti2_data.R
\name{haiti2_data}
\alias{haiti2_data}
\title{Clean haiti cases data for Model 2}
\usage{
haiti2_data()
}
\value{
Data frame
}
\description{
Generate a data frame of department level reported cases of cholera
in Haiti.
}
\examples{
haiti2_data()
}
