% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti2_jesse.R
\name{haiti2_jesse}
\alias{haiti2_jesse}
\title{Build SpatPomp object for Model 2}
\usage{
haiti2_jesse(
  cutoff = 2014.161,
  region = "before",
  measure = "linear",
  joint = FALSE
)
}
\arguments{
\item{cutoff}{a numeric. Used to define the cutoff time for the model}

\item{region}{a character string. Specifies whether the time region is before
or after the cutoff time.}

\item{measure}{a character string. Specifies the measurement model used. Valid
entries are "linear" and "log".}

\item{joint}{a boolean. Specifies if the state values should carry over from
epidemic to endemic phase.}
}
\value{
An object of class \sQuote{spatPomp}.
}
\description{
This is a basic modified copy of haiti2().
Generate a \sQuote{spatPomp} object for fitting to Haiti cholera data. Based on
the model developed by Laura Matrajt et al. at the University of Florida.
}
\examples{
c_epidemic <- haiti2(cutoff=2014.25, region="before")
}
