% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti3_9units.R
\name{haiti3_9units}
\alias{haiti3_9units}
\title{Build pomp object for model 3.}
\usage{
haiti3_9units(dt_years = 1/365.25)
}
\arguments{
\item{dt_years}{step size, in years, for the euler approximation.}
}
\value{
\code{\link[spatPomp]{spatPomp}} representation of model 3 described in \href{https://www.sciencedirect.com/science/article/pii/S2214109X20303107}{Lee, Elizabeth et. al.} and it's accompanying \href{https://ars.els-cdn.com/content/image/1-s2.0-S2214109X20303107-mmc3.pdf}{Supplemental Material}.
}
\description{
This function creates a version of the \sQuote{spatPomp} representation of
Model 3, with only 9 units (excluding the Sud Est department, see the
supplement material for details).
}
\details{
Generate a \sQuote{spatPomp} object for fitting to Haiti cholera data.
This model is a stochastic compartmental model applied at the
level of the ten Haitian departments. It is the stochastic
translation of a deterministic SIRB model based on Ordinary
Differential Equations (ODEs), and has been implemented as a
discrete-state model based on a Partially-Observed Markov Process
(POMP), simulating the stochastic transitions between compartments
as discrete events.

The model subdivides the population of each department into
compartments counting the number of individuals at the different
stages of the disease: Susceptible individuals (S),
Infected symptomatic (I), infected Asymptomatic (A), and
Recoverd (R). The main feature of this model is that it contains
an environmental compartment describing the bacterial concentration (B)
in the local environment, which is used to estimate the force of
infection.

This model removes the redundancy in the Recovered compartments found
in the original code.

This model was developed by Lemaitre, Joseph, et. al at the Laboratory
of Ecohydrology, Ecole Polytechnique Federale de Lausanne (CH).
}
\examples{
mod3 <- haiti3_9units()
}
\seealso{
\code{\link{haiti2}} and \code{\link{haiti1}} for other models used
to fit the cholera epidemic in Haiti.
}
