% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti3_dep_data.R
\name{haiti3_dep_data}
\alias{haiti3_dep_data}
\title{Function used to get department data for haiti3_dep model}
\usage{
haiti3_dep_data(departement = "Artibonite")
}
\description{
Function used to get department data for haiti3_dep model
}
