% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti3_panel.R
\name{haiti3_panel}
\alias{haiti3_panel}
\title{Build PanelPomp representation of model 3. The shared parameters are
set to the best fit on data from Artibonite, and the unit specific
parameters are set to their best independent values using the values
fit in Artibonite.}
\usage{
haiti3_panel(
  delta.t = 1/365,
  departements = c("Artibonite", "Centre", "Grande_Anse", "Nippes", "Nord", "Nord-Est",
    "Nord-Ouest", "Ouest", "Sud", "Sud-Est"),
  betaB_trend = FALSE,
  start_time = "2014-03-01",
  B0 = FALSE,
  old_mod = FALSE
)
}
\arguments{
\item{delta.t}{delta time step used in the model}

\item{departements}{vector of departements to include in the panelPomp.
The default is all departements.}

\item{start_time}{Time of which to start the panel. In the
original paper, the authors used a start time of 2014-03-01 for all
departements except for Ouest, which started with a time of 2017-06-10.
This means that there wasn't a national-coupled model fit to the
data until 2017-06-01, which means that there was less than 2 years
of data to fit the model.}

\item{B0}{Boolean indicator. If TRUE, Initial value of the Bacteria
compartment will be estimated; If FALSE, the equilibrium will be used.}

\item{BetaB_trend}{Boolean indicator of whether or not a trend parameter
should be added to the betaB parameter.}
}
\description{
Build PanelPomp representation of model 3. The shared parameters are
set to the best fit on data from Artibonite, and the unit specific
parameters are set to their best independent values using the values
fit in Artibonite.
}
