% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haitiRainfallData.R
\docType{data}
\name{haitiRainfall}
\alias{haitiRainfall}
\title{Haiti Rainfall Data}
\format{
The dataset contains 6911 rows and 11 columns:
\describe{
\item{date}{Date stored as year, month, day of rainfall observation}
\item{Artibonite:Sud-Est}{These columns contain daily rainfall number for
each of the respective departements.}
}

This dataset was created by loading the .csv file saved as
inst/extdata/rainfall.csv.
}
\usage{
haitiRainfall
}
\description{
This dataset contains daily rainfall data for each departement in Haiti from
2000-03-01 to 2019-01-31. The dataset is used as a covariate in Model 3
}
\keyword{datasets}
