% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelParms_toSpatParms3.R
\name{panelParms_toSpatParms3}
\alias{panelParms_toSpatParms3}
\title{Panel to Spat parameters}
\usage{
panelParms_toSpatParms3(panel_parms)
}
\arguments{
\item{panel_parms}{Any set of parameters for a panel_pomp version of
model 3. Typically, these parameters will be the parameters that resulted
in the best fit for the model.}
}
\description{
This function converts the parameters fit for haiti3_panel to parameters
that work for haiti3_spatPomp. This is done because the parameters are
fit using panel_pomp, but ultimately the coupling requires the use of a
spatPomp object in order to use the block particle filter.
}
