% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_haiti3.R
\name{profile_haiti3}
\alias{profile_haiti3}
\title{Profiling haiti3}
\usage{
profile_haiti3(
  search1 = list(NBPF = 5, NP = 50, SPAT_REGRESSION = 0.5, NREPS = 6, NP_EVAL = 100,
    NREPS_EVAL = 6, RW_SD = NULL, COOLING = 0.5, KEEP_TRACES = FALSE, KEEP_LIK_MAT =
    FALSE),
  search2 = list(NBPF = 5, NP = 50, SPAT_REGRESSION = 0.5, NREPS = 6, NP_EVAL = 100,
    NREPS_EVAL = 6, RW_SD = NULL, COOLING = 0.5, KEEP_TRACES = FALSE, KEEP_LIK_MAT =
    FALSE),
  prof_parameter,
  prof_values,
  nprof,
  start_date = "2010-11-20"
)
}
\arguments{
\item{search1}{search params}

\item{search2}{search params}

\item{prof_parameter}{variable to profile}

\item{prof_values}{values of profile variable}

\item{nprof}{number of profile searches}

\item{start_date}{start date for model.}
}
\value{
The best stuff.
}
\description{
TODO: Before actually running anything, create a "lab-log" of all of the things that we tried, that way
}
