% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_from_filter2.R
\name{project_from_filter2}
\alias{project_from_filter2}
\title{Project from Filtering Distribution}
\usage{
project_from_filter2(mod, PF, covarGen = NULL, nsims = 100, seed = 123)
}
\arguments{
\item{mod:}{POMP model that you would like to simulate from.}

\item{PF:}{pfilter object, the result of calling
pfilter(..., save.states = TRUE). Note that the states must be saved in
this object for the function to work.}

\item{covarGen:}{function that is used to generate covariates, if the model
needs covariates to run rprocess.}

\item{nsims:}{Integer number of simulations desired.}

\item{seed:}{seed for the random number generator.}
}
\description{
This function simulates from a Haiti-cholera POMP model into the future
starting from filtering distribution at the the last time point available.
}
\details{
Because we are primarily interested in projecting the number of Cholera
cases in the future under various vaccination scenarios, it makes sense to
use as much information as possible, as simulating from the start of the
time series will ultimately disregard the known truth at the end of the
available date. The state of the system at the end of the time series
can be accounted for by simulating the filtering distribution:

\deqn{Xk | Y1 = y1*, \ldots, Yk*}

As currently implemented, the efficiency could be greatly improved for
projecting models that do not have covariates.
}
