% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_rain.R
\name{project_rain}
\alias{project_rain}
\title{Project Rainfall}
\usage{
project_rain(end_date = as.Date("2029-12-20"))
}
\arguments{
\item{end_date:}{Date object representing when the projection of rainfall
should stop.}
}
\description{
This function projects rainfall in each of the Haitian departements. The
rainfall is used as a covariate in Model 3.
}
\details{
The rainfall projection is made by assuming that the rainfall on certain
dates in the future will be similar to the rainfall on the same dates in
previous years. Therefore the projection is made by the following steps:
\enumerate{
\item Starting from 1 day after the available data (either actual data
or projected rainfall from previous steps), get the 14-day window of dates
in which rainfall will be projected.
\item With this 14-day window, randomly select a year in the past in which
actual rainfall data is available.
\item Project the rainfall on the selected 14-day window by using the
observed rainfall measurements over the same dates in the randomly
selected year.
\item Repeat the process until a projection for all desired dates in the
future is obtained.
}

Note that this forecasting procedure is stochastic and that a different
rainfall projection will be obtained each time the function is called. While
it is unlikely that this function will provide precise estimates of future
rainfall, it will likely capture seasonal trends in rainfall that may affect
the spread of Cholera in Haiti.
}
