% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vacc_scenario3.R
\name{run_vacc_scenario3}
\alias{run_vacc_scenario3}
\title{Simulate Vaccination Campaign.}
\usage{
run_vacc_scenario3(
  scenario_strs = c("S0", "S1", "S2", "S3", "S4", "S25"),
  nsim = 20
)
}
\arguments{
\item{scenario_strs}{character vector containing the name of each vaccination
scenario that will be simulated. The options are:
TODO: update this list
\itemize{
\item S0: Default vaccination scenario
\item S1
\item S2
\item S3
\item S4
\item S25
}}

\item{nsim}{integer number of simulations for each vaccination scenario.}
}
\description{
Simulate Vaccination Campaign.
}
