% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_h3_fit_parms.R
\name{set_h3_fit_parms}
\alias{set_h3_fit_parms}
\title{Set Model 3 hyper-parameters for model fitting}
\usage{
set_h3_fit_parms(RUN_LEVEL, rw_1.sd, rw_2.sd, rw_3.sd)
}
\arguments{
\item{RUN_LEVEL}{Integer value 1-3. This determines the computational effort
used to fit the model. \code{RUN_LEVEL == 1} is used primarily for debugging
purposes, \code{RUN_LEVEL == 2} uses sufficient computational effort to
obtain preliminary results, and \code{RUN_LEVEL == 3} gives results of the
full computation.}

\item{rw_1.sd}{Random walk standard deviation specification for the first
search.}

\item{rw_2.sd}{Random walk standard deviation specification for the second
search.}

\item{rw_3.sd}{Random walk standard deviation specification for the third
search.}
}
\value{
A list containing hyper-parameters for fitting model 3.
}
\description{
This function sets hyper-parameters needed to fit Model 3. The primary
argument is RUN_LEVEL, which determines the computational effort used to
fit the model. The computation effort is primarily determined by the number
of particles, number of BPF iterations, and the number of replicates.
}
\details{
Originally, this function did not exist and these hyper-parameters were
just set in the R code right before fitting the model. Ultimately this
function was created with the intention of making the code in the ms.Rnw
file that uses this package easier to read.
}
\examples{
rwsd <- rw_sd(epsilon = 0.02)
set_h3_fit_parms(RUN_LEVEL = 3, rw_1.sd = rwsd, rw_2.sd = rwsd, rw_3.sd = rwsd)
}
