% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti1_vacscen.R
\name{vac_scen}
\alias{vac_scen}
\title{Make covariate table for Model 1}
\usage{
vac_scen(scen_code)
}
\arguments{
\item{scen_code}{Code for vaccination scenario}
}
\value{
A list of codes specific to the given vaccination scenario
}
\description{
Generate a data frame containing weekly values for periodic basis splines for the seasonal transmission term beta and vaccination efficacy values
}
\details{
NOTE: # represents the code of a different vaccination deployment strategy
0 = no vaccinations,
1 = fast national,
2 = 2-department,
3 = slow national,
4 = 3-department,
25 = fast national high coverage, and all others represent combinations
}
