% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaccinationScenario3.R
\name{VaccinationScenario3}
\alias{VaccinationScenario3}
\title{Vaccination Scenarios}
\usage{
VaccinationScenario3(
  course_year,
  percent_uv,
  percent_1d,
  percent_2d,
  ve = 1,
  not_dep = c()
)
}
\arguments{
\item{course_year}{Number of years that the vaccination campaign is conducted.}

\item{percent_uv}{Percent of completly unvaccinated individuals.}

\item{percent_1d}{Percent of individuals that received a single dose of vaccination.}

\item{percent_2d}{Percent of individuals that received two doses of the vaccination.}

\item{ve}{\code{c(1, 2, 3)}. Vaccination Efficacy. Only 1 is used by original authors.}

\item{not_dep}{Departments that are not part of the vaccination campaign.}
}
\value{
List containing parameters of the vaccination campaign.
}
\description{
\code{VaccinationScenario3} returns lists containing various vaccination scenarios.
}
