% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti1_ve_decay.R
\name{ve_decay}
\alias{ve_decay}
\title{Vaccination efficacy by month for Model 1}
\format{
A data frame with 61 rows and 41 variables.
\describe{
\item{month}{number of months after vaccination}
\item{ve_s1}{vaccine efficacy rate for assumption set 1}
\item{ve_s2}{vaccine efficacy rate for assumption set 2}
\item{ve_s3}{vaccine efficacy rate for assumption set 3}
}
}
\source{
\url{https://zenodo.org/record/3360991#.YNC2Iy1h0ow}
}
\description{
Dataset containing monthly waning vaccine effectiveness for three vaccine effectiveness assumptions.
}
\references{

}
\concept{datasets}
