% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_coupled_haiti3.R
\name{fit_coupled_haiti3}
\alias{fit_coupled_haiti3}
\title{Fit Model 3 SpatPomp}
\usage{
fit_coupled_haiti3(
  search1 = list(NBPF = 5, NP = 50, SPAT_REGRESSION = 0.5, NREPS = 6, NP_EVAL = 100,
    NREPS_EVAL = 6, RW_SD = NULL, COOLING = 0.5, KEEP_TRACES = FALSE),
  search2 = list(TOP_N = 2, NBPF = 5, NP = 50, SPAT_REGRESSION = 0.5, NREPS = 3, NP_EVAL
    = 100, NREPS_EVAL = 6, RW_SD = NULL, COOLING = 0.5, KEEP_TRACES = FALSE),
  search3 = list(TOP_N = 2, NBPF = 5, NP = 100, SPAT_REGRESSION = 0.5, NREPS = 3, NP_EVAL
    = 100, NREPS_EVAL = 6, RW_SD = NULL, COOLING = 0.5, KEEP_TRACES = FALSE),
  ncores = 3,
  nsearches = 2
)
}
\arguments{
\item{ncores}{Number of cores used to fit the model. The code is written
so that the optimal number of cores with \code{RUN_LEVEL = 3} is 36.}

\item{nsearches}{integer number of searches to conduct. See details below.}

\item{search[i]}{list containing parameters used to fit the model. See details.}
}
\description{
This function fits the spatPomp version of Model 3 to cholera incidence
data from Oct 2010 - Jan 2019 in Haiti. The data are measured at a
weekly timescale.
}
