% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_haiti1.R
\name{fit_haiti1}
\alias{fit_haiti1}
\title{Fit Model 1}
\usage{
fit_haiti1(
  NP = 50,
  NMIF = 3,
  NUM_TREND = 3,
  NPROF = 3,
  NREPS_EVAL = 3,
  NP_EVAL = 50
)
}
\arguments{
\item{NP}{Number of particles to use for the Iterated Filter}

\item{NMIF}{Number of MIF iterations}

\item{NUM_TREND}{Number of points on profile to try for the trend parameter.}

\item{NPROF}{Number of global search for each value of the trend parameter.}

\item{NREPS_EVAL}{Number of particle filter replications to evaluate the
likelihood.}

\item{NP_EVAL}{Number of particles used for likelihood evaluation.}
}
\description{
This function fits Model 1 by performing calculating a Monte-Carlo adjusted
profile for the linear trend in transmission parameter. It is assumed that
a parallel backend compatable with the foreach and doRNG packages has been
registered prior to calling this function
}
\examples{
\dontrun{fit_m1 <- fit_haiti1()}

}
