% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_haiti1.R
\name{fit_haiti1}
\alias{fit_haiti1}
\title{Fit Model 1}
\usage{
fit_haiti1(version = "basic", run_level = 1)
}
\arguments{
\item{version}{In c("basic", "joint").}

\item{run_level}{in c(1, 2, 3)}
}
\description{
This function fits the specified version of model 1.
}
\details{
As of Feb 20, 2022, Model 1 can be written in 4 different forms:
\itemize{
\item Basic: aggregated, single time period \code{pomp} object (\code{haiti1()}).
\item Joint: aggregated, joint time periods \code{pomp} object (\code{haiti1_joint()}).
\item Disagg: disaggregated \code{pomp} object (\code{haiti1_disagg()}).
\item Panel: disaggregated \code{panelPomp} object (\code{haiti1_disagg()}).
}

The Basic version is the closest to the model described and used by Lee et al.
in that it only has an additional overdispersion term. We currently
provide methods for fitting both the Basic and Joint forms.
}
\examples{
m1 <- haiti1_joint()
fit_m1 <- fit_haiti1(version = "joint", run_level = 2)

}
