% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_haiti2.R
\name{fit_haiti2}
\alias{fit_haiti2}
\title{Fit Haiti 2}
\usage{
fit_haiti2(initialization, obj_fun)
}
\arguments{
\item{initialization}{A vector of parameter values that will be used as
the initialization in order to perform optimization. This initialization
must match the necessary parameters in the obj_fun argument.}

\item{obj_fun}{The objective function to maximize, obtained as the output
of \link[pomp:traj_match]{pomp::traj_objfun} of an instance of Model 2.}
}
\value{
A vector containing the estimated parameters and the corresponding
likelihood value.
}
\description{
This simple function takes as input an objective function to minimize and
parameter initializations, and returns the corresponding maximum likelihood
estimate. Note that for this function to work properly, the objective
function needs to be created using \link[pomp:traj_match]{pomp::traj_objfun}.
}
