% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_haiti3.R
\name{fit_haiti3}
\alias{fit_haiti3}
\title{Fit Model 3}
\usage{
fit_haiti3(
  n_searches = 3L,
  search1 = list(NP = 20, NMIF = 3, NREPS = 3, NREPS_EVAL = 3, NP_EVAL = 25),
  search2 = list(TOP_N = 1, NP = 20, NMIF = 3, NREPS = 3, NREPS_EVAL = 3, NP_EVAL = 25),
  search3 = list(TOP_N = 1, NP = 20, NMIF = 3, NREPS = 3, NREPS_EVAL = 3, NP_EVAL = 25),
  search4 = list(TOP_N = 1, NP = 20, NMIF = 3, NREPS = 3, NREPS_EVAL = 3, NP_EVAL = 25),
  search5 = NULL,
  ncores = 3
)
}
\arguments{
\item{n_searches}{integer number of searches to conduct. See details below.}

\item{ncores}{Number of cores used to fit the model. The code is written
so that the optimal number of cores with \code{RUN_LEVEL = 3} is 36.}

\item{search[i]}{list containing parameters used to fit the model. See details.}
}
\description{
This function fits the panelPOMP version of Model 3 to cholera incidence
data from Oct 2010 - Jan 2019 in Haiti. The data are measured at a
weekly timescale.
}
\details{
Because of the complex structure of the model, and the high-dimensional nature
of panelPomp models, Model 3 requires several rounds of Panel Iterated
Filtering in order to obtain the best possible fits. This function breaks the
model fitting process into 6 different phases:
\itemize{
\item{Phase 1: }{Global search of all model parameters.}
\item{Phase 2: }{Local search of all unit specific parameters, with starting
values derived from the top \emph{TOP_N} results of the global search.}
\item{Phases 3-4: }{Local search of all unit-specific parameters, starting
from the top \emph{TOP_N} results of the previous local search.}
\item{Phase 5: }{Local search for all parameters.}
}

The user has some control over each of these phases. Importantly, by setting
\emph{n_searches}, a user may make the fitting process shorter by avoiding
some of the later phases. Additional MIF2 search parameters that can be set
for each phase are the following:
\itemize{
\item{TOP_N: }{Number of results from the previous search to use as a
starting point for the next search. Note that this argument is ignored for
the first global search.}
\item{NP: }{Number of particles to use in the Panel Iterated Filtering.}
\item{NMIF: }{Number of MIF iterations to use in the Panel Iterated Filtering.}
\item{NREPS: }{For each starting point, how many replicated searches should
be conducted. Here, we recommend that NREPS * TOP_N be a multiple of
ncores.}
\item{NREPS_EVAL: }{Number of times to replicate the evaluation using a
particle filter.}
\item{NP_EVAL: }{Number of particles to use in model evaluation.}
\item{RW_SD: }{Specification of the rw.sd used in the MIF2 algorithm.}
}
}
