% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_haiti3.R
\name{fit_haiti3}
\alias{fit_haiti3}
\title{Fit Model 3}
\usage{
fit_haiti3(RUN_LEVEL = 1, ncores = 1)
}
\arguments{
\item{RUN_LEVEL}{parameter in \code{c(1, 2, 3)}. The different RUN_LEVELS
correspond to the computational effort that is used to fit the model, with
\code{RUN_LEVEL = 1} being the least ammount of computation used to fit and
\code{RUN_LEVEL = 3} being the largest ammount of computation used to fit the
model. Note that \code{RUN_LEVEL = 3} is used to obtain our final results, and
\code{RUN_LEVEL = 1} is primarily used for debugging purposes. More precise
information about each of the run levels is provided in the Details
Section below.}

\item{ncores}{Number of cores used to fit the model. The code is written
so that the optimal number of cores with \code{RUN_LEVEL = 3} is 36.}
}
\description{
This function fits model 3 to the available haiti cholera data
}
\details{
As of Feb 16, 2022, Model 3 can be written in 3 different forms:
\itemize{
\item Coupled \code{pomp} object (\code{haiti3_correct()}).
\item Independent \code{panelPomp} object (\code{haiti3_panel()}).
\item Coupled \code{spatPomp} object (\code{haiti3_spatPomp()}).
}

Despite being represented in so many different forms, the model is only fit
using \code{panelPomp} methods. This is because it best replicates what was done
in Lee et al. by not using \code{spatPomp} (which was not available at the time).
The modeling fitting is improved from the fitting done by Lee et al by using
\code{panelPomp}, which came out just before the Lee et al. paper did.

Because of the complex nature of the \code{panelPomp} model, we found that
performing the model fitting in three different stages (1. Performing a
global search of all model parameters, 2. Fix the unit-specific parameters
found in the global search, and perform another global search just on the
shared parameters, 3. Fix the shared parameters found in the global search
and perform a local search on the unit-specific parameters). Note that adding
more additional stages or and changing any of the three stages we used may
have resulted in a slightly different estimate.

The following is a more in depth description of the \code{RUN_LEVEL} argument:
\describe{
\item{\code{RUN_LEVEL = 1}}{Stage 1: \code{Np = 50} and \code{Nmif = 3} for global search,
with 3 starting locations for the parameters used. Stage 2: \code{Np = 50} and
\code{Nmif = 3} and 3 starting locations for global search. Stage 3: \code{Np = 50}
and \code{Nmif = 3}, repeating the local search 3 times. At the end of each
stage, 50 particles are used to estimate the likelihood, and 3 replications
of the estimation are conducted.}
\item{\code{RUN_LEVEL = 2}}{Stage 1: \code{Np = 400} and \code{Nmif = 15} for global search,
with 10 starting locations for the parameters used. Stage 2: \code{Np = 400} and
\code{Nmif = 15} and 10 starting locations for global search. Stage 3: \code{Np = 400}
and \code{Nmif = 15}, repeating the local search 3 times. At the end of each
stage, 400 particles are used to estimate the likelihood, and 5 replications
of the estimation are conducted.}
\item{\code{RUN_LEVEL = 3}}{Stage 1: \code{Np = 1000} and \code{Nmif = 50} for global search,
with 72 starting locations for the parameters used. Stage 2: \code{Np = 1000} and
\code{Nmif = 50} and 72 starting locations for global search. Stage 3: \code{Np = 1500}
and \code{Nmif = 50}, repeating the local search 36 times. At the end of each
stage, 36 replications are conducted with \verb{Np = 2000, 2500, 3000} at stages
1, 2, and 3, respectively}
}
}
