% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti1.R
\name{haiti1}
\alias{haiti1}
\title{Build pomp object for Model 1}
\usage{
haiti1()
}
\value{
An object of class \sQuote{pomp}.
}
\description{
Generate a class \sQuote{pomp} object for fitting to epidemic/endemic Haiti cholera data.
}
\examples{
m1 <- haiti1()
}
