% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti1_joint.R
\name{haiti1_joint}
\alias{haiti1_joint}
\title{Build pomp object for Model 1 - Joint}
\usage{
haiti1_joint(vacscen = "id0")
}
\arguments{
\item{vacscen}{ID code for vaccination scenario}
}
\value{
An object of class \sQuote{pomp}
}
\description{
Generate a class \sQuote{pomp} object for fitting to epidemic/endemic Haiti cholera data jointly.
}
\examples{
m1 <- haiti1_joint(vacscen = "id0")
}
