% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti2_vaccScenario.R
\name{haiti2_vaccScenario}
\alias{haiti2_vaccScenario}
\title{Model 2: Project Vaccination Scenario}
\usage{
haiti2_vaccScenario(epi_params, end_params)
}
\arguments{
\item{epi_params:}{a vector of parameters for the epidemic period}

\item{end_params:}{a vector of parameters for the endemic period}
}
\value{
List containing two objects: \describe{
\item{tot_inc}{Vector estimating the total cholera incidence from Feb 2019-Feb 2024 for each scenario.}
\item{all_trajs}{data.frame containing trajectories for each of the vaccination scenarios.}
}
}
\description{
This function projects the following vaccination scenarios under Model 2:
\describe{
\item{V0}{The default vaccination scenario: no additional vaccinations are deployed.}
\item{V1}{Two department (Centre and Artibonite) over two years.}
\item{V2}{Three department (Centre, Artibonite, and Ouest) over two years.}
\item{V3}{Countrywide vaccination over 5 years.}
\item{V4}{Countrywide vaccination over 2 years.}
}
}
\examples{
h2 <- fit_haiti2()
results <- haiti2_vaccScenario(epi_params = h2$epi_params, end_params = h2$end_params)

}
