% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haiti3_dep.R
\name{haiti3_dep}
\alias{haiti3_dep}
\title{Build pomp object for individual departments following model 3}
\usage{
haiti3_dep(
  departement = "Artibonite",
  start_time = "2014-03-01",
  delta.t = 1/365.25 * 0.2,
  betaB_trend = FALSE,
  B0 = FALSE
)
}
\arguments{
\item{departement}{String indicating which departement to fit the
pomp model to.}

\item{start_time}{Time of which to start the time series. In the
original paper, the authors used a start time of 2014-03-01 for all
departements except for Ouest, which started with a time of 2017-06-10.
This means that there wasn't a national-coupled model fit to the
data until 2017-06-01, which means that there was less than 2 years
of data to fit the model. Still, the authors fit the remaining departemental
models to data from 2014-03-01. This added argument allows us to
fit the Ouest model earlier, or fit the remaining departement models
at 2017-06-01}

\item{delta.t}{Delta time step used in Euler's approximation}

\item{betaB_trend}{Whether or not to include a trend parameter for
betaB.}

\item{B0}{Boolean indicator. If TRUE, Initial value of the Bacteria
compartment will be estimated; If FALSE, the equilibrium will be used.}
}
\value{
\code{\link[pomp]{pomp}} object.

\code{\link[pomp]{pomp}} representation of model 3 described in \href{https://www.sciencedirect.com/science/article/pii/S2214109X20303107}{Lee, Elizabeth et. al.} and it's accompanying \href{https://ars.els-cdn.com/content/image/1-s2.0-S2214109X20303107-mmc3.pdf}{Supplemental Material}.
}
\description{
Build pomp object for individual departments following model 3
}
